refactor slatepack definition into separate module
